Resolve #2230 -  DeepSSM Augmentation should write VTK files in addition to particles
